+++ servicetalk-http-api/src/main/java/io/servicetalk/http/api/HttpServerBuilder.java
@@ -42,7 +42,7 @@ import java.util.Map;
 import java.util.function.BooleanSupplier;
 import java.util.function.Predicate;
 
-import static io.servicetalk.concurrent.api.internal.BlockingUtils.blockingInvocation;
+import static io.servicetalk.concurrent.api.BlockingUtils.blockingInvocation;
 
 /**
  * A builder for building HTTP Servers.
